AI/webcrawling/requirements.txt
selenium
Pillow
requests
webdriver-manager
beautifulsoup4